Refactor: Centralize Bitunix kline fetching logic in apiUtils
